third_party/blink/web_tests/external/wpt/css/css-animations/parsing/keyframes-name-invalid-expected.txt

This is a testharness.js-based test.
[FAIL] invalid: @keyframes "" { }
  assert_equals: expected 0 but got 1
Harness: the test ran to completion.
